chore(mcp): split bridge client tests
